+++ frameworks/asacfwk/src/accessibility_element_operator_async_mng.cpp
@@ -29,7 +29,7 @@ bool AccessibilityElementAsyncOperatorMng::SearchElementResultTimer(const int se
     std::lock_guard<std::recursive_mutex> lock(mutex_);
     struct timeval getTime {};
     gettimeofday(&getTime, NULL);
-    uint64_t startTime = getTime.tv_sec * SECOND_TO_MILLIS + getTime.tv_usec;
+    uint64_t startTime = (uint64_t)(getTime.tv_sec * SECOND_TO_MILLIS + getTime.tv_usec);
     HILOG_DEBUG("element sequence[%{public}d]", sequence);
 
     do {
@@ -37,8 +37,8 @@ bool AccessibilityElementAsyncOperatorMng::SearchElementResultTimer(const int se
             return true;
         }
         gettimeofday(&getTime, NULL);
-        uint64_t endTime =  getTime.tv_sec * SECOND_TO_MILLIS + getTime.tv_usec;
-        uint64_t waitTime = endTime - startTime;
+        uint64_t endTime = (uint64_t)(getTime.tv_sec * SECOND_TO_MILLIS + getTime.tv_usec);
+        uint64_t waitTime = (uint64_t)(endTime - startTime);
         if (TIMEOUT_OPERATOR_MILLIS < waitTime) {
             completed_ = true;
             HILOG_ERROR("Failed to wait sequence[%{public}d], feedbackSequence_[%{public}d]",
